+++ CHANGELOG.md
@@ -38,6 +38,8 @@ The following breaking changes only affect you if you're using named arguments:
 - `BigInteger::modPow()` now uses `$exponent` and `$modulus` as parameter names
 - `BigInteger::shiftedLeft()` now uses `$bits` as the parameter name
 - `BigInteger::shiftedRight()` now uses `$bits` as the parameter name
+- `BigInteger::isBitSet()` now uses `$bitIndex` as the parameter name
+- `BigInteger::randomBits()` now uses `$bitCount` as the parameter name
 - `BigDecimal::withPointMovedLeft()` now uses `$places` as the parameter name
 - `BigDecimal::withPointMovedRight()` now uses `$places` as the parameter name